Created function parsePosition
Takes the comma-dilimited position coordinates in cells and converts them to a "#rows"x3 array of doubles with one col per coordinate (x in 1 col, y in 1 col and z in 1 col). 

created to format monleyLogic data such that it resembles old monkeylab data and can be processed with same functions. 

Run it with: 
parsedStruct = parsePosition(initialStruct);

Where:  
initialStruct = logicData(2).UEData.UE_Position